FinSight-Agents/requirements.txt
Flask
google-cloud-bigquery
google-cloud-aiplatform
pandas
numpy
requests
beautifulsoup4
plotly
scikit-learn
pytest
jupyter
adk
openai
python-dotenv